AT client and cellular change: add UART power saving.
All the u-blox modules supported by ubxlib support "UART power saving" (AT+UPSV) which allows the module to go into a slow-clocked mode, not driving the UART HW, after a certain amount of inactivity on the UART interface (and of course the radio interface).  The cellular module is woken up through either RTS line activity or a character being sent to it on the UART (depending on the configuration and the module) or by a need to send something out over the UART interface (e.g. a URC).  In the case of what the system integration manuals call the "Wake-up via data reception" mechanism, that first wake-up character is lost and hence a wake-up call-back has been added to the AT client to bring the module to life before normal AT operation can resume.

This commit switches on UART power saving for all modules using this latter "Wake-up via data reception" mechanism (but see note below), applying an inactivity timeout of 6 seconds (to match that of the SARA-R4 cellular modules on which this timeout is not configurable).  You should notice no difference in operation aside from reduced power consumption and an extra "AT->OK", possibly "AT->AT->OK" sequence in the AT client logs every now and then which is the module being woken-up.  If you wish to remove the feature simply add the conditional compilation flag U_CFG_CELL_DISABLE_UART_POWER_SAVING to your build.

Note that this is NOT "3GPP power saving" (AT+CPSMS), which will be implemented in a future commit, but is a pre-requisite for it (at least in the case of the SARA-R5 cellular module).

Note: this commit will NOT switch on power saving for non-SARA-R4 modules that have their flow control lines connected to the MCU.  The reason for this is that when the module goes to sleep the CTS line from the module floats high ("off") and so the MCU's UART HW will refuse to send anything to it, locking us out.  The fix for this is to use the RTS line to the module as the sleep signaller if it is connected but this is non-trivial as the RTS line CANNOT then be used for flow control (since no UART HW will set the RTS pin to "on" sufficiently far in advance of transmitting data for the module to wake up in time).  Support for this mode of operation will be added in a later commit.  SARA-R4 modules appear to hold the CTS line low even while asleep (and do not, in any case, support use of the RTS pin as a sleep-signalling mechanism), hence they can support this kind of sleep even with the flow control lines connected/used in the normal way.

Test: *